Implement and exercise the GSheets interface
There is now a logical, compartmentalised chain of functionality that
supports writing a dataframe to a google sheets hosted spreadsheet.

There was some regression here from messing about with merges
principally because the sheet model, facade and adapter responsible for
working with the cloud backend wasn't being tested.

Tests have been added to properly fix this functionality and test the
new stuff.