Make sure types coerced early (and this is tested)
There was some inconsistency in ensuring values sourced from a google
sheet are the correct type. This ensures types are changed in the
GSheetCODA class when constructing the three dataframes representing the
three "sub" tables (requirements, characteristics and relationships
between them).

This is tested by making sure the dataframe we're using contains strings
which is how we are using pygsheets currently. Using numerisation in
pygsheets (or possibly the underlying google api client) isn't having
the intended effect and it's better to just coerce ourselves.