Remove trailing columns in CODA source sheet
GSheetCODA now constructs a dataframe itself via the get_rows method on
GSheetsFacade rather than using pygsheets to create a dataframe.
pygsheets provides no out-of-the-box way of stripping blank trailing
columns as far as I know, and we've already done this for binary
weighting matrices in the WorksheetAdapter's get_all_values method which
get_rows uses.