fix: compatibility issue with newer pandas
df.iloc throws a fit if you pass in an array of booleans. This is a
little unintuitive because it's just the reverse of what we do to
extract the columns of interest and manipulate them before inserting
them back in.

There may be a better way to do all of this generally