fix: pandas no longer supports out of bounds usecols
Previously when parsing an xlsx CODA sheet, we just set to the max
column to CZ and chomped up as many characteristics that were defined.

Pandas no longer likes this, so take a different approach to parsing the
sheet - don't specify the columns; grab everything then slice the
dataframe.

This is a backwards compatible fix.